+++ packages/core/src/utils/cursor.ts
@@ -1,5 +1,24 @@
 import type { AnyColumn, AnyColumnWithTable, OperandValueExpression, SelectQueryBuilder } from "kysely"
-import type { Cursor, Pageable } from "./types"
+import { z } from "zod"
+
+export interface PaginatedResult<T> {
+  data: T[]
+  next: Cursor | null
+}
+
+// Cursors are opaque strings that are parsed in the individual repositories
+export const CursorSchema = z.string()
+
+export const PaginateInputSchema = z
+  .object({
+    take: z.number(),
+    cursor: CursorSchema.optional().nullable(),
+  })
+  .optional()
+  .default({ take: 20, cursor: undefined })
+
+export type Pageable = z.infer<typeof PaginateInputSchema>
+export type Cursor = z.infer<typeof CursorSchema>
 
 type GetNextCursorOptions<T> = {
   pageable: Pageable
@@ -20,10 +39,10 @@ function getNextCursor<T>(records: T[], options: GetNextCursorOptions<T>): Curso
 
 type PaginatedQueryOptions<DB, TB extends keyof DB, C extends AnyColumn<DB, TB> | AnyColumnWithTable<DB, TB>, O> = {
   pageable: Pageable
-  decodeCursor: (cursor: Cursor) => OperandValueExpression<DB, TB, C>
-  buildCursor: (record: O) => Cursor
-  column: C
   order: "asc" | "desc"
+  decodeCursor?: (cursor: Cursor) => OperandValueExpression<DB, TB, C>
+  buildCursor?: (record: O) => Cursor
+  column: C
 }
 
 /**
@@ -44,7 +63,6 @@ export async function singleColPaginatedQuery<
 >(query: SelectQueryBuilder<DB, TB, O>, options: PaginatedQueryOptions<DB, TB, C, O>) {
   const {
     pageable: { cursor, take },
-    decodeCursor,
     order,
   } = options
 
@@ -55,12 +73,24 @@ export async function singleColPaginatedQuery<
     }
   }
 
+  if (options.column === "id") {
+    return singleColPaginatedQuery(query, {
+      ...options,
+      decodeCursor: decodeUlidIdCursor as (cursor: Cursor) => OperandValueExpression<DB, TB, C>,
+      buildCursor: buildUlidIdCursor as (record: O) => Cursor,
+    })
+  }
+
+  if (options.decodeCursor === undefined || options.buildCursor === undefined) {
+    throw new Error("decodeCursor and buildCursor must be provided when column is not 'id'")
+  }
+
   // Take N+1 to determine if there is a record behind `take`.
   const pageWithNextLength = take + 1
   let pagedQuery = query.limit(pageWithNextLength)
   pagedQuery = pagedQuery.orderBy(options.column, order)
   if (cursor !== undefined) {
-    const decodedCursor = decodeCursor(cursor)
+    const decodedCursor = options.decodeCursor(cursor)
     pagedQuery = pagedQuery.where(options.column, order === "asc" ? ">=" : "<=", decodedCursor)
   }
 
@@ -79,4 +109,23 @@ export async function singleColPaginatedQuery<
     next: nextCursor,
     data: records.slice(0, take), // Don't include take+1 record used for determining next page
   }
+}
+
+export const base64Encode = (value: string) => Buffer.from(value).toString("base64")
+export const base64Decode = (value: string) => Buffer.from(value, "base64").toString("ascii")
+
+type RecordWithUlidId = { id: string }
+export function buildUlidIdCursor<O extends RecordWithUlidId>(record: O): Cursor {
+  return base64Encode(record.id)
+}
+export function decodeUlidIdCursor(cursor: Cursor): string {
+  return base64Decode(cursor)
+}
+
+type RecordWithCreatedAt = { createdAt: string }
+export function buildCreatedAtCursor<O extends RecordWithCreatedAt>(record: O): Cursor {
+  return base64Encode(record.createdAt)
+}
+export function decodeCreatedAtCursor(cursor: Cursor): string {
+  return base64Decode(cursor)
 }
